launcher, converter: Introduce CPUDomainConfigurator
Partially extract the logic responsible for configuring the CPU to a
dedicated `CPUDomainConfigurator`.
