+++ pkg/virt-launcher/virtwrap/converter/compute/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "clock.go",
         "console.go",
         "controllers.go",
+        "cpu.go",
         "graphics.go",
         "host_device.go",
         "hypervisor.go",
@@ -38,6 +39,7 @@ go_library(
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/topology:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
+        "//pkg/virt-launcher/virtwrap/converter/vcpu:go_default_library",
         "//pkg/virt-launcher/virtwrap/launchsecurity:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
